Move peer connection on restart into the AdminService
Before, members in existing circuit would be created in the
splinterd start function. With the integration of the
PeerManager this functionality is moved into the AdminService
because the AdminService must be able to hold on to the
associated peer_refs.

Also adds adding peers for pending proposals on restart.

Finally renames restart_services to re_initialize_circuits,

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>