PeerManager retry interval configurable
Makes the retry_interval used by the Pacemaker
configurable so that tests do not hang too long.

Finally fixes a log statement in Pacemaker if the sender
shuts down first. This was being caused by the Pacemaker
checking if its running and then sleeping. It should sleep
and then check.

Also signals PeerManager and ConnectionManager to shutdown in
tests before waiting for one to complete.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>