Complete authorization on receipt of AUTHORIZE
This change require tracking the states more closely. The state of
remote acceptance and remote identified have been added, so that these
states can be properly transitioned through to the final Authorized
state.

It also allows for the use of the state value to indicate whether or not
authorization is complete, as well as hold the identity.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>