Enhance log message when frame protocol version is incorrect
Prior to this change, a generic log message was printed with very little
useful information. After this change, the local endpoint and frame
version information is printed in the log message, making it much
clearer what happened. After this change, it is also possible to tell
the difference between endpoints in the log message, which is useful
when binding to multiple endpoints.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>