Remove `service_endpoint` from CLI status response
Remove the `service_endpoint` field from the `NodeStatus` struct in the
CLI, which is used to parse the response from a `GET /status` request
made to splinterd. This field was recently moved behind an experimental
feature, so it is not included in the response by default. This field is
not currently required by the CLI, so it can be removed.

Signed-off-by: Logan Seeley <seeley@bitwise.io>