Override graceful shutdown when reconnect is enabled
Setting reconnect to true will override graceful shutdowns.
Previously, a server sending any none abnormal close reason (i.e.
"away") would cause the reconnect process to skip. This is not always
desirable, so we are updating the WebSocketClient to override this
functionality when reconnect is set to true. When reconnect is set to
true the WebSocket connection can only be terminated by the client.

Signed-off-by: Ryan Banks <rbanks@bitwise.io>