Guard deprecated Network structs with #[cfg(test)]
These structs were deprecated with the integration of the
PeerManager/ConnectionManager. However, they are still widely
used in tests.

This commits guards the structs to only be available in tests,
and only public from within the crate thus removing them from
the public API of splinter.

Also moves the old Network structs from mod.rs to network.rs
in an attempt to make it clear what is deprecated.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>