Update incoming error handling for Gameroom requests
Updates the error handling on the response to a request made to
gameroomd, in that the original error is sent along to the function
so that the correct error message may be displayed to the user,
based on the status of the response within the original error. This
also moves a more generic error message to be logged to the console
immediately.

This also updates the GameDetail and GameroomDetail pages to handle
the different possible response status codes to better reflect to the
user the error that has occurred. If different status codes are
encountered, the correct error page is now loaded rather than
a 404-Not-Found page for all errors caught, as was previously
implemented.

Signed-off-by: Shannyn Telander <telander@bitwise.io>