Check correlation id before logging
Adds a check for if the correlation id is not present in an admin
message and circuit direct message, and if it is, an additional space
is added to the end of the correlation id before it is added to the log
message. This ensures there is not an extra space in the case that it
is not present.

Signed-off-by: Shannyn Telander <telander@bitwise.io>