Add Subcommand or help setting to all subcommands
Add the SubcommandRequireOrHelp setting to all top-level subcommands
that did not have the setting.  This will display help text if the
command does not include any of the child subcommands.

For example, prior to this commit

    $ splinter admin

would display an error.  After this commit, it will display the help
text for the admin subcommand.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>