Add health service rest resources
If the experimental feature "health" is enabled, add the REST resources
for the service.

This requires a small bit of Option-juggling, as there needs to be a
valid variable that lives beyond the creation block of the health
service, into the block where these threads are joined. The join handle
struct for the health service is put into an option, if the feature is
enabled, and expects to be there during the join phase.

In the case where the feature is disabled, this is just an Option<()>,
which is explicitly dropped during the join phase, so as to not require
any unused-variable exceptions.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>