Update Handler tests to use MockSender
This commit updates the test for Handlers to use a MockSender
that will keep track of the Peer_ID the message is being sent
to and the message in bytes.

This allows for the removal of the Network object and for these
tests to be run without startup up Mesh or requiring inproc
connections.

The assertion for peer_id is switched from expected sender to
expected recipient

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>