Add PeerMangerBuilder for PeerManager
Adds a builder that will start the PeerManager when
calling PeerManagerBuilder.start(). The builder will
call PeerManager::build() which is a private constructor
that will start the background thread. The PeerManager
new and start functions are now deprecated. This
commit also adds a new function to get
PeerManagerConnectors.

The tests and splinterd have been updated to use the new builder
and a test has been added to verify back word compatibility.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>