Update subscribing to PeerManagerNotfications to take a sender
This commit deprecates subscribe() in favor of subscribe_sender
which takes a Sender<T> which can be used to get a
PeerManagerNotfication and adds the ability to convert the
notification to another type. This change updates the PeerManager
to match the ConnectionManager.

Also adds an unsubscribe option to PeerManagerConnector so a
subscriber can unsubscribe from notifications.

Also removes Clone and PartialEq from PeerManager Messages.
These decorators were not required and caused issues with the
adding of the new subscriber messages.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>