Add new transport::tls:: module
This module replaces the previously deprecated module and contains TLS
code that is not implementation specific (it could, for example be used
between both socket and websocket implementations).

Includes TlsConfig which encapsulates the configuration information we
expect for our transport implementations.

Also includes functions to create SslConnectors and SslAcceptors from
TlsConfig.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>