Remove ConnectionPool in favor of r2d2
Removes the `libsplinter::database::ConnectionPool` struct and its
associated code, and updates Diesel store implementations to use the
`r2d2` library for connection pools.

Also updates a doc test to use an in-memory store implementation instead
of a DB store implementation (because it's simpler) and disables the doc
test to avoid compilation errors that depend on which features are
enabled.

Signed-off-by: Logan Seeley <seeley@bitwise.io>