Replace ConnectionError with MigrationError
The `libsplinter::database::error::ConnectionError` was only used in one
place (the biome migrations) and the error does not make sense for this
usage. It is replaced with a more general `MigrationError` in the biome
migrations module. The old error is now aliased to the `MigrationError`,
compiled behind the `biome` feature (that is the only place it's used),
and deprecated.

Also updates some comments and a log statement to more accurately
reflect the purpose of the migrations; they are not specific to the
users or credentials tables, but rather for all biome tables in a
PostgreSQL database.

Signed-off-by: Logan Seeley <seeley@bitwise.io>