Use gameroom circuit template in Gameroom daemon
Updates the `propose_gameroom` REST API endpoint function to use a
CircuitCreateTemplate to generate the CreateCircuit admin message
used to propose a gameroom.

This also updates the module for ApplicationMetadata to add a feature
flag over two ApplicationMetadata methods and an error variant that
are now only used behind the `test-authorization-handler` feature.

Signed-off-by: Shannyn Telander <telander@bitwise.io>