Log error in case file is not readable
Changes the `list_available_templates` function to log an error instead
of return an error in case a file is not able to be read. As this
function intends to just list the templates, it should simply not log
that file and move on as there may be other available templates.

Signed-off-by: Shannyn Telander <telander@bitwise.io>