Add store module to libsplinter
Adds the `libsplinter::store` module, which is behind the
`store-factory` experimental feature. This module contains a new
`StoreFactory` trait, which is an abstract factory for building stores
backed by a single storage mechanism (e.g. database). Currently this
trait allows for building all of the Biome stores. PostgreSQL and SQLite
implementations of the `StoreFactory` trait are provided. This module
also includes a `create_store_factory` function for easily constructing
a store factory for a given storage location.

Signed-off-by: Logan Seeley <seeley@bitwise.io>