Add encryption cost arg for updating credentials
Adds the `password_encryption_cost` as a new argument to the
`CredentialsStore::update_credentials` and
`KeyStore::update_keys_and_password` methods. This argument is
required to allow configuration of the encryption level for passwords
when updating credentials. Before this change, the update methods just
used the default encryption cost, which may not always be the desired
behavior.

Includes an update to the in-memory credentials and key stores to
encrypt the passwords; previously, these stores did not actually encrypt
the passwords on update.

Also fixes a typo of the word "encryption" in a variable name.

Signed-off-by: Logan Seeley <seeley@bitwise.io>