Remove "a:" and "r:" prefix to template arguments
Takes away the "a:" and "r:" prefix to template argument names. This was
previously used as a special namespace to identify the template
arguments based on the section of the template they came from. However,
this has been deemed unnecessary and somewhat confusing as is, so the
prefixes are being removed.

Signed-off-by: Shannyn Telander <telander@bitwise.io>