Change "into_builders" to "apply_to_builder"
Changes the `CircuitTemplate` `into_builders` method to
`apply_to_builder`, which takes in a `CreateCircuitBuilder` and applies
the template rules using the builder .with_* methods. This builder is
then returned. Also removes the `Builders` struct as this is no longer
necessary.

Also changes the CLI representation of `CircuitTemplate` `into_builders`
method into `apply_to_builder`, which takes a mutable reference to a
`CreateCircuitMessageBuilder` and adds the template's resulting
`CreateCircuitBuilder` to this builder object.

Signed-off-by: Shannyn Telander <telander@bitwise.io>