Add AdminServiceStore trait and data structs
This PR defines an API for managing the API for writing and
reading circuit state and pending circuit proposals. The
provided `AdminServiceStore` trait makes no assumptions about
the storage backend.

This module also include the structs that will be stored/returned by
the service store.

This is added behind an experimental feature as it is not currently used.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>