Add YAML backed AdminServiceStore implementation
This commit provides a implementation of the AdminServiceStore
that supports the existing (v0.4 compatible) state YAML files.

The store will take in paths to the existing circuit state file and
proposal state file and load existing state into the store. If the
files do not exist, empty state files are created. As circuits and
proposals are added to state, the store will write the updated
yaml files to state.

Note the service definition in the existing yaml file differs from
the definition used in the admin service store, as such they
must be converted when read/written.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>