Update CLI builder module to handle templates
Updates the CreateCircuitMessageBuilder module in the CLI to more
accurately handle an instance of a circuit template. Now, this checks
the existing `CreateCircuitBuilder` object for any values that have
already been filled in before overwriting these values with default
values. This ensures that any value from the template applied to the
`CreateCircuitBuilder` is not erased, but also that the values submitted
through CLI options are given precedence over the template values.

Signed-off-by: Shannyn Telander <telander@bitwise.io>