Remove error propagation from listener thread
In the case of add_inbound_connection errors, these should be logged
immediately, instead of being propagated out of the thread closure.
Propagating errors in this way result in the error effectively being
swallowed until the application is exited, and then only if the result
returned from the thread join is logged.

The thread exiting behavior remains the same.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>