Rework `apply_rules` method
Reworks the `apply_rules` method so the organization makes it clear
which `rule` is creating the value and that the value is being added to
the CreateCircuitBuilder using a `with_*` method.

Now, all individual `rules` modules have an `apply_rule` method which
returns the value to be added to the CreateCircuitBuilder. The
`apply_rules` method then uses `with_*` statements on the
CreateCircuitBuilder to add these values.

Signed-off-by: Shannyn Telander <telander@bitwise.io>