Prepare AdminServiceStore for diesel backend
Initializes a struct for the AdminServiceStore diesel backend, the
DieselAdminServiceStore. This struct holds the connection to the
database. Also implements the Clone trait for this struct with a
Sqlite connection and PostgreSQL connection to enable this struct to be
implemented for each type of database.

Also defines the AdminServiceStoreOperations struct which will be used
to contain the operations used to implement the AdminServiceStore trait.

Signed-off-by: Shannyn Telander <telander@bitwise.io>