Add AdminServiceStore trait implementation
Adds the AdminServiceStore trait implementation for the
DieselAdminServiceStore, which is the representation for the database
backend. This uses the operations to implement the AdminServiceStore
methods for both the SQLite and PostgreSQL connection types.

Separate implementations are required for the SQLite and PostgreSQL
backends as the SQLite connection does not support Diesel's
`SupportsDefaultKeyword` trait, which is required when using a generic
backend type.

Signed-off-by: Shannyn Telander <telander@bitwise.io>