Support multiple directories in template storage
Adds a SPLINTER_CIRCUIT_TEMPLATE_PATH environment variable for optional
paths to template storage, which also necessitates support within the circuit
template module to search multiple directories for template files. While
loading a circuit template, directories are searched from a list which
includes the environment variable, if set, and the default location if
the template name passed in is not an absolute or relative path.
Also updates the CLI template commands to check for this environment
variable, as well as the default template directory.

Also adds the glob dependency for the circuit-template feature to allow
for template files to be collected across multiple possible directories.

Signed-off-by: Shannyn Telander <telander@bitwise.io>