Move CircuitProposal and builder into own file
Moves CircuitProposal into
store::circuit_proposal::{CircuitProposal, CircuitProposalBuilder}

Structs are still exposed only in store:: via pub use. This change
also requires getters to be implemented on the structs to be used
throughout the store implementations.

This commit also adds a builder for VoteRecord and removes the pub
from the fields in VoteRecord and CircuitProposal.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>