Remove serde serilization decorators from AdminServiceStore struct API
Serde should not be a part of the public API of the structs for the
AdminServiceStore.

This commit also add Yaml specific structs for CircuitProposal,
ProposedCircuit, and VoteRecord because the Vec<u8> in those
structs must be converted to a hex string for storage. This
was done automatically before with the serde decorators, and now
needs to be done explicitly.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>