Update YamlAdminServiceStore to write to temp files and then rename
Before it would be possible to get an IOError partially through writing
out the state and for the file to become corrupted as a result.
Instead, this update the store to  write out the file then
rename the file to its final location.

fs::rename corresponds to the rename function on Unix and the MoveFileEx
function with the MOVEFILE_REPLACE_EXISTING flag on Windows.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>