Replace allowed_nodes with node_id in Service
The AdminServiceStore's Service will not include
allowed_nodes (a list of node ID's the service could connected to)
and will instead only have node_id, the node ID of the specific
node the service will connect to. One node is already enforced by
the admin service during circuit request validation.

Converting the allowed_nodes to node id is still required for
backword compatibility for protobuf messages, as well as
YamlAdminServiceStore.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>