Add OAuth Provider struct
Adds the `Provider` struct to the `oauth` module. This struct is an
abstraction of an OAuth provider that wraps an OAuth client (provided
by the `oauth` library) and the scopes that will be used to create
authentication requests.

This commit only includes a `new` method for configuring the provider;
future commits will add more methods for interacting with the provider.

Signed-off-by: Logan Seeley <seeley@bitwise.io>