Add OAuth provider configuration to REST API
Adds the `oauth_provider` field to the `RestApi` struct, and a
corresponding field to the REST API's builder. This field is not yet
used, but it provides the configuration needed for implementing
authentication later on.

This configuration is guarded by the new `oauth` experimental feature.
This commit also adds logic to the REST API's builder that requires some
kind of authentication be configured when the `auth` feature is enabled.
Currently, `oauth` is the one authentication type that can be
configured, but others may be added in the future.

Also updates the REST API's doc test to be compiled but not run because
of the complexity of the auth configuration depending on which features
are enabled. This complexity would lessen the value that the doc test
provides. All other tests that use the REST API builder have been
updated to include a mock OAuth provider when the `oauth` feature is
enabled.

Signed-off-by: Logan Seeley <seeley@bitwise.io>