Remove foreign key to circuit member in node_endpoint
The foreign key was using node_id, however node_id is not
unique in circuit_members and the circuit_member primary key is
(circuit_id, node_id). belongs_to does not support
composite foreign key.

Since a node is not specific to a circuit, we already manually
check if we should delete node endpoints when circuit
members are deleted.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>