Update middleware wrapping in Actix REST API
Updates the `RestApi::run` method to apply middleware with distinct
assignments instead of a chain. This requires re-declaring the `app`
variable when each middleware is added because the generic type of the
`App` is changed with each new middleware that's added.

This change is intended to make the addition of future middleware
easier, especially when the new middleware is behind a feature;
conditionally compiling the chain of middleware with multiple features
would be messy and require a lot of duplication.

Also updates the `Logger` middleware to be the outermost wrapper so all
requests are logged.

Signed-off-by: Logan Seeley <seeley@bitwise.io>