Make OAuthResourceProvider crate public
Updates the `OAuthResourceProvider` to be crate public rather than
public outside of the crate. This resource provider is not used directly
by the consumer of the splinter crate, but created by the Splinter REST
API itself. Making this provider public leads the consumer of the crate
to believe that they need to add the endpoints themselves, which is not
true.

Signed-off-by: Logan Seeley <seeley@bitwise.io>