Add authorization middleware components
Adds the `Authorization` middleware for the Actix REST API that checks
all requests to determine the client's identity. This middleware will be
used to enforce authentication by resolving the Authorization header to
an identity using configured identity providers and enable access
control (in a future commit).

The new `splinter::auth::identity::IdentityProvider` trait is used to
fetch a client's or user's identity based on some authentication.
This trait will be implemented for different auth types and auth
providers. Splinter initially supports OAuth identity providers, but
more types will likely be added in the future.

Signed-off-by: Logan Seeley <seeley@bitwise.io>