Add traits and data structs for Biome-oauth
This change adds a linkage between an OAuth user id and a biome user-id.
The linkage is achieved through the use of the OAuthUser and its
associated store.

It will enable support of signing key management on behalf of a user
that has authenticated via an OAuth provider.

This commit includes only the structs/enums for the data representations
of an OAuth user, the trait for the OAuthUserStore, and the errors that
may occur during any of the store's operations.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>