Add crate::error::InternalError
Adds the crate::error module and the first library-level error,
InternalError. This error module will provide common reusable errors for
the library, reducing redundancy and inconsistency.

InternalError is useful for sitautions where: a) the caller is not at
fault for the error; and b) no hint to the caller on how to handle the
error is desirable.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>