Remove `auth` from splinterd experimental features
When the `auth` feature is enabled, REST API authorization must be
configured to start splinterd. Because OAuth is currently the only
supported auth type, this is not easy to workaround in development or
test environments. Until some basic auth type is added that can be used
for these scenarios when running with experimental features, the `auth`
feature won't be part of the experimental group of features.

Signed-off-by: Logan Seeley <seeley@bitwise.io>