Replace the use of sawtooth-sdk with transact-sdk-javascript
This commit updates the gameroom-app to use the transact sdk
instead of the sawtooth sdk.

This dependency is currently only available through github so
the docker image has to be changed to install as a non root
user.

This will also fix a build issue with the current gameroom-app
build.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>