Prepend OAuth access token with type string
Prepends the access token returned to the browser by the OAuth callback
endpoint with the string `OAuth2:`. By prefixing the token with the
type, the client can treat the access token provided by the Splinter
REST API as an opaque bearer token; it does not have to attach the type
itself.

Signed-off-by: Logan Seeley <seeley@bitwise.io>