Update list circuits and proposal to be sorted by circuit_id
The order proposals and circuits are returned should be deterministic
because they will be used by the rest-api, and paging will require that
they are returned in the same order during multiple calls.

To accomplish this, set order() on the queries and remove the HashMap
around the circuit and proposals so the order returned matches the order
fetched.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>