Replace use of authorize with User extension
When the "auth" feature is enabled, replace the use of authorize with
the User extension in the biome endpoints that may be usable regardless
of authorization scheme.

Under this feature, these endpoints require an authorization scheme to
be in place to function appropriately.

When the "auth" feature is not enabled, this uses the stable biome
method of parsing the token at request time.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>