Refactor `UserTokens` struct to `UserInfo`
Refactors the `UserTokens` to include the user identity, requiring the
name change to `UserInfo` as this struct now encapsulates more information
than just tokens. The `OAuthClient` now gets the user's `identity` from the
OAuth provider immediately after the tokens are retrieved. As a result, this
also includes refactoring the `SaveTokensOperation` to `SaveUserInfoOperation`
as this trait now takes the `UserInfo` struct, not just `UserTokens`.
The `TokenSaveConfig` enum has also been renamed to represent this
struct, `UserInfoSaveConfig`.

This also affects the `oauth/callback` endpoint as the `identity` is now
included in the query string generated from the `UserInfo` returned by
the `OAuthClient`. The `identity` is added to the query string as
`display_name`.

Signed-off-by: Shannyn Telander <telander@bitwise.io>