Move postgres feature to stable in CLI
This moves postgres to stable, which was already true because database
requires it.

This also removes the link between database and postgres, because the
old approach wouldn't work when we add sqlite support.

The CLI migrate command still requires enabling both database and
postgres or compile errors will occur - this will be resolved when
sqlite support is added to the code.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>