Represent access tokens with specific enum
Adds an enum, `AccessToken`, to represent access tokens in the
OAuthUserStore to allow for the `access_token` field to be represented
by the database as a nullable string. This allows for the access token
field to be `Authorized` with the token string, or `Unauthorized` with
no token string. A user may be `Unauthorized` after logging out, for
example.

Signed-off-by: Shannyn Telander <telander@bitwise.io>