Configure CORS middleware once
This change alters the configuration of the CORS middleware to be before
the HttpServer instance is created. This instance is passed in and
cloned in the HttpServer's closure -  this closure is called per thread
created for the server, resulting in a CORS log message per thread.
After this change, there is only a single message logged.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>