Replace manual prefix strip with strip_prefix
This change fixes an issue flagged by Clippy (as of Rust 1.48.0) by
replacing manual prefix stripping (i.e. starts_with combined with a
slice) with the use of the standard function strip_prefix.

Why is this bad (from the clippy documentation)

Using str:strip_{prefix,suffix} is safer and may have better performance
as there is no slicing which may panic and the compiler does not need to
insert this panic code. It is also sometimes more readable as it removes
the need for duplicating or storing the pattern used by
str::{starts,ends}_with and in the slicing.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>