Add `/oauth/logout` endpoint
Adds an `/oauth/logout` endpoint to the Splinter REST API that removes
an `OAuthUser`'s access and refresh tokens. Also updates the API spec to
include this endpoint.

This also updates the `authorize` function to explicitly only skip the
`oauth/login` and `oauth/callback` endpoints and to check the
`oauth/logout` endpoint, rather than skip authorizing it as it is
prefixed with `/oauth`.

Signed-off-by: Shannyn Telander <telander@bitwise.io>