Add experimental sqlite support to splinter database migrate
If sqlite is enabled, the default database connection string
is now /var/lib/splinter/splinter_state.db. This commit
also adds support for changing this location to
SPLINTER_HOME/data/splinter_state.db and
STATE_DIR/splinter_state.db.

If the connection string is :memory: the sqlite migrations
will be run against an in memory sqlite database. This can
be used as "dry-run" test for sqlite migrations.

If sqlite is not enabled, the old default is returned:
postgres://admin:admin@localhost:5432/splinterd

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>