Add CustomizeConnection to turn on foreign key support for sqlite
foreign key support is not turned on by default in SQLite. It can be
turned on by running:

`PRAGMA foreign_keys = ON;`

However this needs to be run on connection creation and using a migration
does not keep it on.

Since we are using a Pool to get Sqlite connections, we need to define a
CustomizeConnection that will turn foreign key support on for the
connection when the connection is required.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>