Replace OAuthClient constructors with Builder
This change replaces the OAuthClient constructors with builders. A
general builder maps to the generic constructor, and Github-specific and
OpenID-specific builders for those configurations.

It also introduces an error module, as the builder requires an enum
error for the three types of errors that may occur at construction time.

The original constructor has been made private, and `new_github` and
`new_openid` removed entirely.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>