Add NewOAuthUserAccess token for store insert
This commit changes the OAuthUserStore add_oauth_user function to take a
NewOAuthUserAccess struct.  This struct has slightly different
constraints then updates or the retrieved value from the store, in that
it must have an access token.

Additionally, it removes the transformation of the OAuthUserAccess into
a model object in the operation, but makes the add_oauth_user operation
only rely on the model objects as input.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>