Add `auth` to `ScabbardClient` and builder
Add an `auth` field to the `ScabbardClient` struct. This field will hold
the string to be submitted to the Splinter REST API in an authorization
header. This field is behind the experimental feature client-auth.

Add a `with_auth` function to the `ScabbardClientBuilder` to set the
`auth` field of the struct.

Modify tests to use the `with_auth` function when building the client
if the client_auth feature is enabled.

Modify requests to the Splinter REST API to submit the string stored in
the `auth` field in an authorization header when the client-auth feature
is enabled.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>