Alter oauth_user to remove UNIQUE constraint
This change alters the oauth_user table, in both postgres and sqlite, to
remove the uniqueness constraints on user_id and provider_user_ref. This
will allow for multiple entries based on access token.

Additionally, the migrations add a last_updated column, which may be
used to cleanup old logins after a set period of time.  The cleanup is
left as a future exercise.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>