Update admin REST API to handle new protocol
The circuit and proposal routes will now check the version of
the SplinterProtocolVersion in the header of the request and
will return the correct version.

An unsupported version will never be handled in the routes because
they will have been caught by the guard.

Also updates the version requested by gameroom and the cli to be
version 1, as this is what is currently supported.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>