Configure InflightOauthRequestStore in splinterd
This change updates splinterd to configure the InflightOauthRequestStore
via the store factory.  It requires reorganizing where and how the
store-factory is created, so that the factory is available for
additional components beyond biome, as well as maintains that the same
factory is used across all store usage.

The one exception is the backwards-compatible memory value for the admin
store.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>