Update ADMIN_SERVICE_PROTOCOL_VERSION and handle backwards compat
Update ADMIN_SERVICE_PROTOCOL_VERSION to version 2. If the
AdminService have agreed on protocol 1, any circuits that
include display name will be rejected.

This is required because 0.4 admin services (speaking protocol 1)
do not support display name and consensus will fail do to differing
bytes/hashes.

If display_name is not set, the bytes are the same.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>