Add OAuthUserIdentityProvider
Adds an initial implementation of the OAuthUserIdentityProvider. The
identity provider will be responsible for periodically re-authenticating
OAuth users as needed. This implementation is not complete; it always
uses the identity from the session store, and it uses the OAuth subject
rather than the Biome ID. Re-authentication and returning the Biome ID
will be implemented in a future commit.

This commit includes updating the Splinter REST API to use this identity
provider for checking identities in the authorization middleware.

Signed-off-by: Logan Seeley <seeley@bitwise.io>