Add experimental https-bind REST API configuration
This change adds the ability to bind using HTTPS, under the experimental
feature "https-bind".  The RestApi struct can be configured with a cert
and a key.

It includes a From implementation for the RestApiError for the
openssl::error::ErrorStack, which may occur during the configuration of
the RestApi struct when https is enabled.

Additionally,

- Corrects many tests to use port 0 instead of manually scanning for
  open ports.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>