Add traits to implement event store database
Adds the necessary traits to implement the operations for the
AdminServiceEventStore, as well as the initial structure for the
operations module. The traits added includes the
`DieselAdminServiceEventStore` and the `AdminServiceEventStoreOperations`
traits.

The Clone implementations for the `DieselAdminServiceEventStore` are
also added, behind an `AdminServiceEventStore`-specific experimental
feature representing the type of database used, PostgreSQL or SQLite.

Signed-off-by: Shannyn Telander <telander@bitwise.io>