Use OAuth client for getting subject identifiers
Updates the `OAuthClient` to provide a new `get_subject` method that
calls its internal `SubjectProvider`. This removes the need for any
other code to directly deal with a subject provider.

This change is motivated by the fact that the
`OAuthUserIdentityProvider` will need to have access to the
`OAuthClient` to be able to use refresh tokens to get new access tokens.
Because this is the only code that uses subject providers other than the
OAuth client and the identity provider needs to access the client
anyway, it makes sense to have the identity provider just reference the
client for all questions it has about an OAuth user.

Signed-off-by: Logan Seeley <seeley@bitwise.io>