Update OAuth id provider to use refresh token
Updates the `OAuthUserIdentityProvider` to use an OAuth refresh token
(if there is one) when re-authentication fails because the access token
is unauthorized.

This includes the addition of the `exchange_refresh_token` method to the
`OAuthClient` struct.

Unit tests have not yet been added for this functionality, since they
require mocking out an OAuth server. This will be addressed with a
future change.

Signed-off-by: Logan Seeley <seeley@bitwise.io>