Add conversions for event store database models
Adds conversions for the models used by the database-backend of
the `AdminServiceEventStore`, from and into the native representations.

Adds `list_from_proposal_with_id` methods for the event store database
models to allow for easily extracting the objects from a
`messages::CircuitProposal`.

Also adds a TryFrom implementation for the event store's
`AdminServiceEvent` to be converted from the database model that
represents this struct.

Signed-off-by: Shannyn Telander <telander@bitwise.io>