Update REST API config to have Azure + Google
Updates the REST API OAuth config to handle Azure and Google separately
from generic OpenID OAuth providers. Refresh tokens are enabled
differently between these two providers, so it's necessary for the REST
API to know which is being used so it's able to correctly request a
refresh token when getting getting a user's access token for the first
time.

This change includes updating splinterd configuration for this change,
along with the splinterd man page.

Signed-off-by: Logan Seeley <seeley@bitwise.io>