Convert admin messages to admin store structs
Adds conversions for the admin store `CircuitProposal` and associated
structs from the admin messages circuit proposal (and related structs)
definitions.

The admin messages `SplinterService` definition includes an
`allowed_nodes` field, which requires `TryFrom` in order to convert this
field to the admin store's `ProposedService` equivalent `node_id` field.
This requires `TryFrom` implementations for all structs that hold the
`ProposedService`s, hence the implementation of `TryFrom` for the
`ProposedCircuit` and `CircuitProposal`.

Signed-off-by: Shannyn Telander <telander@bitwise.io>