Use `AdminServiceEvent` in event store
Uses the `AdminServiceEvent` defined in admin::service::event for the
`AdminServiceEventStore`, changing the return types of the event store
and updating the `AdminServiceEventStore`'s database models to include
conversions from the event store models to the `AdminServiceStore`'s
representation.

Previously, the event store returned the `AdminServiceEvent` defined in
admin::service::messages, which needed to be updated to the
admin::service::event `AdminServiceEvent` to ensure the `CircuitProposal`
representation matches the rest of the admin service.

This also removes redundant conversions that are already implemented for the
`AdminServiceStore`, keeping conversions necessary to convert the
`messages::AdminServiceEvent` objects.

Signed-off-by: Shannyn Telander <telander@bitwise.io>