Add permissions to REST API endpoints
Updates the Splinter REST API's resource definitions to include a
permission for all endpoints. These permissions will be used by the REST
API to determine whether a client is permitted to make a request. The
permission is only present when the `authorization` feature is enabled.

This change includes updating all existing endpoints to provide a
permission.

Also changes doc tests for the `splinter::rest_api` module to be
ignored, since they now require conditional compilation.

Signed-off-by: Logan Seeley <seeley@bitwise.io>